+++ _Typings/editorApi/types/itemComponents/light.d.ts
@@ -2,7 +2,7 @@
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
 declare module "editor:type" {
-    import { IntRange } from "utils"
+    import { NonNegIntRange } from "utils"
 
     enum LightType {
         Point = 0,
@@ -90,7 +90,7 @@ declare module "editor:type" {
         /**
          * @range `1 ~ 256`
          */
-        Quality: IntRange<1, 256>
+        Quality: NonNegIntRange<1, 256>
         StepSize: float
         MaxStep: int
         Jitter: float