+++ _Typings/utils/utils.d.ts
@@ -14,7 +14,10 @@ declare module "utils" {
         ? T
         : TupleToExactLength<L, [...T, T["length"]]>
 
-    type IntRange<L extends int, H extends int> = TupleToExactLength<H> extends [...TupleToExactLength<L>, ...infer R]
+    type NonNegIntRange<L extends int, H extends int> = TupleToExactLength<H> extends [
+        ...TupleToExactLength<L>,
+        ...infer R
+    ]
         ? [...R, H][number]
         : never
 
@@ -24,5 +27,58 @@ declare module "utils" {
         ? C
         : Tuple<T, L, [...C, T]>
 
-    type AssertInt<T extends number> = `${T}` extends `${string}.${string}` ? never : T
+    type ValidateIntString<T extends string> = T extends `${string}e+${string}`
+        ? true
+        : T extends `${string}e-${string}`
+        ? false
+        : T extends `${string}.${string}`
+        ? false
+        : true
+
+    type AssertInt<T extends number> = number extends T ? T : ValidateIntString<`${T}`> extends true ? T : never
+
+    type AssertNeg<T extends number> = number extends T ? T : `${T}` extends `-${string}` ? T : never
+
+    type AssertNonNeg<T extends number> = number extends T ? T : `${T}` extends `-${string}` ? never : T
+
+    type AssertNegInt<T extends number> = number extends T
+        ? T
+        : `${T}` extends `-${infer R}`
+        ? ValidateIntString<R> extends true
+            ? T
+            : never
+        : never
+
+    type AssertNonNegInt<T extends number> = number extends T ? T : `${T}` extends `-${string}` ? never : AssertInt<T>
+
+    type Validate0To1<T extends number> = number extends T
+        ? true
+        : T extends 0
+        ? true
+        : T extends 1
+        ? true
+        : `${T}` extends `-${string}`
+        ? false
+        : `${T}` extends `${string}e+${string}`
+        ? false
+        : `${T}` extends `${string}e-${string}`
+        ? true
+        : `${T}` extends `${infer I}.${string}`
+        ? I extends "0"
+            ? true
+            : false
+        : false
+
+    type Validate0To1s<T extends readonly number[]> = T extends readonly [
+        infer F extends number,
+        ...infer R extends readonly number[]
+    ]
+        ? Validate0To1<F> extends false
+            ? false
+            : Validate0To1s<R>
+        : true
+
+    type Assert0To1<T extends number> = Validate0To1<T> extends true ? T : never
+
+    type Assert0To1s<T extends readonly number[]> = Validate0To1s<T> extends true ? T : never
 }